requirements.txt
Flask==2.0.3
python-dotenv==0.19.2
transformers==4.6.1
torch==1.8.1
